% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_git.R
\name{check_github_credentials}
\alias{check_github_credentials}
\title{check_github_credentials}
\usage{
check_github_credentials(remote)
}
\arguments{
\item{remote}{}
}
\value{
remote
}
\description{
Perform checks to ensure github credentials are valid
found in auth_git.R
}
\examples{
remote <- check_github_credentials(remote)
}
