% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_copy_for_ghqc.R
\name{check_if_qc_file_untracked}
\alias{check_if_qc_file_untracked}
\title{check_if_qc_file_untracked}
\usage{
check_if_qc_file_untracked(qc_file_path)
}
\arguments{
\item{qc_file_path}{path to qc file to be checked for tracking}
}
\value{
boolean if qc file is untracked
}
\description{
found in error_copy_for_ghqc.R
}
\examples{
untracked_file <- check_if_qc_file_untracked(qc_file_path)
}
