% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_issue.R
\name{create_checklists}
\alias{create_checklists}
\title{create_checklists}
\usage{
create_checklists(yaml_path)
}
\arguments{
\item{yaml_path}{path to yaml. Typically within dependent information repo/package}
}
\value{
list of issues created by create_issues()
}
\description{
test with "test_yamls/checklist.yaml"
found in multi_issue.R
}
\examples{
create_checklists(yaml_path)
}
