% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_fix.R
\name{create_comment_body}
\alias{create_comment_body}
\title{create_comment_body}
\usage{
create_comment_body(
  owner,
  repo,
  issue_number,
  message = NULL,
  diff = FALSE,
  reference_commit = "original",
  comparator_commit = "current"
)
}
\arguments{
\item{owner}{owner of repo}

\item{repo}{repo owner}

\item{issue_number}{gh issue number}

\item{message}{message to be added to comment body (default NULL)}

\item{diff}{boolean if difference wanted in body or not (default FALSE)}

\item{reference_commit}{commit to be referenced against (default "original")}

\item{comparator_commit}{commit to be compared (default "current")}
}
\value{
character string of the comment's body
}
\description{
found in qc_fix.R
}
\examples{
comment_body <- create_comment_body(owner, repo, issue_number)
}
