% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_issue.R
\name{create_issue}
\alias{create_issue}
\title{create_issue}
\usage{
create_issue(file, issue_params)
}
\arguments{
\item{file}{file to be QCed}

\item{issue_params}{parameters for the issue to be created}
}
\value{
list with issue number and assignee
}
\description{
create an issue for each file given in a yaml file
\itemize{
\item organize issues associated with a set of files with milestones
\item assign a different user to each issue for a given file
found in multi_issue.R
}
}
\examples{
issue <- create_issue(file, issue_params)

}
