% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_issue.R
\name{create_issues}
\alias{create_issues}
\title{create_issues}
\usage{
create_issues(data)
}
\arguments{
\item{data}{list containing params such as owner, repo, title, description}
}
\value{
list of issues created by create_issue()
}
\description{
create list of issue_params to input to api call
found in multi_issue.R
}
\examples{
issue_params <- create_issues(data)
}
