% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_linux_platform}
\alias{format_linux_platform}
\title{Find and format the linux platform string for ghqc/rpkg installation}
\usage{
format_linux_platform()
}
\value{
the string of the linux platform based on os-release file
}
\description{
Find and format the linux platform string for ghqc/rpkg installation
}
