% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_yaml.R
\name{get_checklists}
\alias{get_checklists}
\title{get_checklists}
\usage{
get_checklists()
}
\value{
checklists_data
}
\description{
read in checklists
found in create_yaml
}
\examples{
checklists_data <- get_checklists()
}
