% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_org_attributes.R
\name{get_current_repo}
\alias{get_current_repo}
\title{get_current_repo}
\usage{
get_current_repo(remote = get_remote())
}
\arguments{
\item{remote}{typically result of get_remote()}
}
\value{
current remote
}
\description{
found in get_org_attributes.R
}
\examples{
current_remote <- get_current_remote(get_remote())
}
