% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_git.R
\name{get_gh_api_url}
\alias{get_gh_api_url}
\title{get_gh_api_url}
\usage{
get_gh_api_url(remote_url)
}
\arguments{
\item{remote_url}{url for remote repo}
}
\value{
configured api url
}
\description{
found in auth_git.R
}
\examples{
gh_api_url <- get_gh_api_url(remote_url)
}
