% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_git.R
\name{get_gh_url}
\alias{get_gh_url}
\title{get_gh_url}
\usage{
get_gh_url(remote_url)
}
\arguments{
\item{remote_url}{}
}
\value{
env_url the environement url which has been checked to that it matches the remote
}
\description{
found in auth_git.R
}
\examples{
env_url <- get_gh_url(remote_url)
}
