% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_org_attributes.R
\name{get_remote}
\alias{get_remote}
\title{get_remote}
\usage{
get_remote(remote_list)
}
\arguments{
\item{remote_list}{list of remotes}
}
\value{
remote
}
\description{
found in get_org_attributes.R
}
\examples{
remote <- get_remote(remote_list)
}
