% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{ghqc_archive_app}
\alias{ghqc_archive_app}
\title{Archive file(s)}
\usage{
ghqc_archive_app(
  app_name = "ghqc_archive_app",
  qc_dir = getwd(),
  lib_path = ghqc_libpath(),
  config_path = ghqc_config_path()
)
}
\arguments{
\item{app_name}{the name of the app to run in the background}

\item{qc_dir}{the directory in which the app is run}

\item{lib_path}{the path to the ghqc package and its dependencies}

\item{config_path}{the path to the ghqc configuring information}
}
\description{
Archive file(s)
}
\seealso{
\link{ghqc_assign_app} and \link{ghqc_record_app}
}
