% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ghqc_config_path}
\alias{ghqc_config_path}
\title{The default install location for the ghqc custom configuration repository}
\usage{
ghqc_config_path()
}
\value{
string containing the default path to the ghqc custom configuration repository (~/.local/share/ghqc/<config repo name here>
}
\description{
The default install location for the ghqc custom configuration repository
}
