% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_01_create.R
\name{ghqc_create_app}
\alias{ghqc_create_app}
\title{ghqc_create_app
found in app_01_create.R}
\usage{
ghqc_create_app()
}
\value{
NA
}
\description{
ghqc_create_app
found in app_01_create.R
}
\examples{
ghqc_create_app()
}
