% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_setup.R
\name{ghqc_example_setup}
\alias{ghqc_example_setup}
\title{Function to set up the ghqc environment, including writing to the .Renviron, custom configuration repository download, ghqc.app dependency installation, and ghqc.app installation if available, for use of the ghqc application suite}
\usage{
ghqc_example_setup(
  config_repo = "https://github.com/A2-ai/ghqc.example_config_repo"
)
}
\arguments{
\item{config_repo}{the URL for the custom configuration repository from which to import organization specific items like checklist templates}
}
\description{
Function to set up the ghqc environment, including writing to the .Renviron, custom configuration repository download, ghqc.app dependency installation, and ghqc.app installation if available, for use of the ghqc application suite
}
