% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ghqc_infopath}
\alias{ghqc_infopath}
\title{The default install location for the ghqc customizing information repository}
\usage{
ghqc_infopath()
}
\value{
string containing the default path to the ghqc information repository
}
\description{
The default install location for the ghqc customizing information repository
}
\examples{
ghqc_infopath()
}
