% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{ghqc_record_app}
\alias{ghqc_record_app}
\title{Generate a QC Record for one or more Milestones}
\usage{
ghqc_record_app(
  app_name = "ghqc_record_app",
  qc_dir = getwd(),
  lib_path = ghqc_libpath(),
  info_path = ghqc_config_path()
)
}
\arguments{
\item{app_name}{the name of the app to run in the background}

\item{qc_dir}{the directory in which the app is run}

\item{lib_path}{the path to the ghqc package and its dependencies}

\item{info_path}{the path to the ghqc configuring information}
}
\description{
This function allows the user to generate a QC Record for one or more Milestones created with \link{ghqc_assign_app}.

\strong{To Generate a QC Record:}
\enumerate{
\item Select one or more Milestones.
\itemize{
\item optional to include both open and closed Milestones by unchecking "Closed Milestones only".
}
\item Optional: input a name for the PDF.
\itemize{
\item The default name is a hyphenated combination of the GitHub repository name and selected Milestone name(s).
}
\item Optional: input the directory in which to generate the PDF.
\itemize{
\item The default directory is the root of the R project.
}
\item Optional: indicate if the report should only include the Milestone and Issue summary tables by checking "Just tables".
Else, the default setting will generate a Record that contains the summary tables up front as well as detailed descriptions
for each Issue including version control information, users, datetimes, events, actions, comments and more.
\item Create the PDF by clicking "Generate QC Record" at the bottom of the pane.
}
}
\seealso{
\link{ghqc_assign_app} and \link{ghqc_resolve_app}
}
