% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{ghqc_report}
\alias{ghqc_report}
\title{ghqc_report}
\usage{
ghqc_report(
  milestone_names = NULL,
  input_name = NULL,
  just_tables = FALSE,
  location = ".",
  owner = get_organization(),
  repo = get_current_repo()
)
}
\arguments{
\item{milestone_names}{names of milestones (default NULL)}

\item{input_name}{name of pdf to be saved (default NULL to set with normal convention)}

\item{just_tables}{boolean to have report be only tables (default FALSE)}

\item{location}{location to save pdf (default to "." for current directory)}

\item{owner}{owner of repo (default to get_organization())}

\item{repo}{repo name (default to get_current_repo())}
}
\value{
NA
}
\description{
ghqc_report
}
\examples{
ghqc_report()
}
