% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_03_report.R
\name{ghqc_report_app}
\alias{ghqc_report_app}
\title{ghqc_report_app
found in app_03_report.R}
\usage{
ghqc_report_app()
}
\value{
NA
}
\description{
ghqc_report_app
found in app_03_report.R
}
\examples{
ghqc_report_app()
}
