% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_setup.R
\name{ghqc_setup}
\alias{ghqc_setup}
\title{Interactive function to set up the ghqc environment, including writing to the .Renviron, custom configuration repository download, and ghqc.app dependency installation/linking, for use of the ghqc application suite}
\usage{
ghqc_setup()
}
\value{
This function is used for its effects, not outputs. It will return the results of any ghqc dependency installation as needed
}
\description{
Interactive function to set up the ghqc environment, including writing to the .Renviron, custom configuration repository download, and ghqc.app dependency installation/linking, for use of the ghqc application suite
}
