% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_logger.R
\name{ghqc_toggle_logger}
\alias{ghqc_toggle_logger}
\title{ghqc_toggle_logger}
\usage{
ghqc_toggle_logger()
}
\value{
NA
}
\description{
toggles amount of logging displayed to console
found in init_logger.R
}
\examples{
ghqc_toggle_logger()
}
