% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_02_update.R
\name{ghqc_update_app}
\alias{ghqc_update_app}
\title{ghqc_update_app
found in app_02_update.R}
\usage{
ghqc_update_app()
}
\value{
NA
}
\description{
ghqc_update_app
found in app_02_update.R
}
\examples{
ghqc_update_app()
}
