% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_packages.R
\name{install_ghqcapp_dependencies}
\alias{install_ghqcapp_dependencies}
\title{install ghqc.app's dependencies into an isolated library}
\usage{
install_ghqcapp_dependencies(
  lib_path = ghqc_libpath(),
  pkgs = ghqc_depends,
  use_pak = TRUE
)
}
\arguments{
\item{lib_path}{\emph{(optional)} the path to install the dependencies. If not set, defaults to "~/.local/share/ghqc/rpkgs"}

\item{pkgs}{\emph{(optional)} list of packages to install. Defaults to ghqc and all of its dependencies}

\item{use_pak}{\emph{(optional)} optionally removes the requirement to have \code{pak} installed in the project repository. Setting to \code{FALSE} will reduce performance}
}
\description{
install ghqc.app's dependencies into an isolated library
}
