% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_shiny_ready}
\alias{is_shiny_ready}
\title{Shiny is "ready" if the download.file is able to serve the starting html,
at this point, we will try to hit the shiny app and see if it downloads}
\usage{
is_shiny_ready(url)
}
\arguments{
\item{url}{the http port}
}
\value{
true or false depending on if Shiny is ready
}
\description{
Shiny is "ready" if the download.file is able to serve the starting html,
at this point, we will try to hit the shiny app and see if it downloads
}
