% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_depends.R
\name{link_ghqcapp_dependencies}
\alias{link_ghqcapp_dependencies}
\title{symlink previously installed package library containing all ghqc.app dependencies to an isolated package library}
\usage{
link_ghqcapp_dependencies(link_path, lib_path = ghqc_libpath())
}
\arguments{
\item{link_path}{the path to the installed package library}

\item{lib_path}{\emph{(optional)} the path to install the dependencies. If not set, defaults to ghqc_libpath()}
}
\value{
this function is primarly used for its effects, but will the results of the symlink
}
\description{
symlink previously installed package library containing all ghqc.app dependencies to an isolated package library
}
