% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_org_attributes.R
\name{list_milestones}
\alias{list_milestones}
\title{list_milestones}
\usage{
list_milestones(org, repo)
}
\arguments{
\item{org}{organization repo lives in}

\item{repo}{repo name}
}
\value{
list of milestones
}
\description{
found in get_org_attributes.R
}
\examples{
list_of_milstones <- list_milestones(org, repo)
}
