% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_client_info.R
\name{load_client_info}
\alias{load_client_info}
\title{load_client_info}
\usage{
load_client_info()
}
\value{
path which client info was cloned to
}
\description{
found in git_client_info.R
}
\examples{
client_repo_path <- load_client_info()
}
