% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_fix.R
\name{post_comment}
\alias{post_comment}
\title{post_comment}
\usage{
post_comment(owner, repo, issue_number, body)
}
\arguments{
\item{owner}{owner of repo}

\item{repo}{repo name}

\item{issue_number}{gh issue number}

\item{body}{body content to be posted into GH}
}
\value{
comment posted to gh
}
\description{
found in qc_fix.R
}
\examples{
comment <- post_comment(owner, repo, issue_number, body)
}
