% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_repo.R
\name{remove_ghqc_configuration}
\alias{remove_ghqc_configuration}
\title{Remove the downloaded customizing information repository from \code{info_path}}
\usage{
remove_ghqc_configuration(info_path = ghqc_infopath())
}
\arguments{
\item{info_path}{\emph{(optional)} path in which the repository, set in environmental variable \code{GHQC_INFO_REPO}, is, or should be, downloaded to. Defaults to \verb{~/.local/share/ghqc/\{repo_name\}}}
}
\value{
this function is used for its effects, but will return the removed \code{info_path}
}
\description{
Remove the downloaded customizing information repository from \code{info_path}
}
