% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_packages.R
\name{remove_ghqcapp_dependencies}
\alias{remove_ghqcapp_dependencies}
\title{Remove all content in the specified lib path. Optionally removes the cache as well.}
\usage{
remove_ghqcapp_dependencies(
  lib_path = ghqc_libpath(),
  cache = FALSE,
  .only_base = FALSE
)
}
\arguments{
\item{lib_path}{\emph{(optional)} the path to the installed dependency packages. If not set, defaults to ghqc_libpath()}

\item{cache}{\emph{(optional)} flag of whether to clear the cache or not. Defaults to keeping the cache}

\item{.only_base}{\emph{(optional)} flag to only delete ghqc.app dependencies in the basepath}
}
\value{
information related to deleted lib path
}
\description{
Remove all content in the specified lib path. Optionally removes the cache as well.
}
