% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_renviron.R
\name{setup_ghqc_renviron}
\alias{setup_ghqc_renviron}
\title{helper function to setup/write Renviron file for ghqc}
\usage{
setup_ghqc_renviron(GHQC_CONFIG_REPO)
}
\arguments{
\item{GHQC_CONFIG_REPO}{Repository URL to the custom configuration repository}
}
\value{
This function is used primarly to write to the ~/.Renviron file. It will return the text contained in ~/.Renviron
}
\description{
helper function to setup/write Renviron file for ghqc
}
