% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_repo.R
\name{download_ghqc_configuration}
\alias{download_ghqc_configuration}
\title{Download the customizing information repository as set in environmental variable \code{GHQC_INFO_REPO}}
\usage{
download_ghqc_configuration(info_path = ghqc_infopath())
}
\arguments{
\item{info_path}{\emph{(optional)} path in which the repository, set in environmental variable \code{GHQC_INFO_REPO}, is, or should be, downloaded to. Defaults to \verb{~/.local/share/ghqc/\{repo_name\}}}
}
\description{
Download the customizing information repository as set in environmental variable \code{GHQC_INFO_REPO}
}
