% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_repo.R
\name{download_ghqc_configuration}
\alias{download_ghqc_configuration}
\title{Download the custom configuration repository as set in environmental variable \code{GHQC_CONFIG_REPO}}
\usage{
download_ghqc_configuration(config_path = ghqc_config_path(), .force = FALSE)
}
\arguments{
\item{config_path}{\emph{(optional)} path in which the repository, set in environmental variable \code{GHQC_CONFIG_REPO}, is, or should be, downloaded to. Defaults to \verb{~/.local/share/ghqc/\{repo_name\}}}

\item{.force}{\emph{(optional)} option to force a new download of the ghqc custom configuration repository}
}
\value{
this function is used for its effects, but will return if bool if the result is a config repo found locally at config_path
}
\description{
Download the custom configuration repository as set in environmental variable \code{GHQC_CONFIG_REPO}
}
