% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ghqc_libpath}
\alias{ghqc_libpath}
\title{The default install location for the ghqc package and its dependencies. If it does not exist, it will be created.}
\usage{
ghqc_libpath()
}
\value{
string containing the default lib path for the ghqc package and its dependencies depending on the user's platform, R version, and os arch:
\verb{~/.local/share/ghqc/rpkgs/<platform>/<R version>/<os arch>}
}
\description{
The default install location for the ghqc package and its dependencies. If it does not exist, it will be created.
}
