% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep.R
\name{ghqc_sitrep}
\alias{ghqc_sitrep}
\title{Situation report for ghqc set-up}
\usage{
ghqc_sitrep(..., lib_path = ghqc_libpath(), config_path = ghqc_config_path())
}
\arguments{
\item{...}{options to expand output. Current option is only "pkgs" to expand list of dependencies}

\item{lib_path}{the path to the ghqc package and its dependencies}

\item{config_path}{the path to the ghqc custom configuration}
}
\value{
This function is primarily used for its printed output, not a returned output
}
\description{
Situation report for ghqc set-up
}
